Phase 2/backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-backend-app"
version = "1.0.0"
description = "Todo backend with authentication"
authors = [{name = "Todo App Team"}]
requires-python = ">=3.7"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.16",
    "SQLAlchemy>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "database*"]